Fix build script

+++ Web/prebuild.ps1
@@ -1,5 +1,8 @@
 cd "..\Hexo"
+yarn
 hexo clean
 hexo generate
-Remove-Item -Recurse -Force "..\Web\wwwroot"
+if (Test-Path "..\Web\wwwroot") {
+  Remove-Item -Recurse -Force "..\Web\wwwroot"
+}
 Copy-Item -Recurse ".\public" "..\Web\wwwroot"
